Выполнено задание 4-1
